refactor(daemon-config): rename IUserConfigurationProvider to ISessionConfigurationProvider
