+++ src/daemon/Tiempitod.NET/Tiempitod.NET.csproj
@@ -29,5 +29,6 @@
 
     <ItemGroup>
       <Folder Include="bin\Linux\net8.0\" />
+      <Folder Include="Configuration\User\" />
     </ItemGroup>
 </Project>
